add: input filtering
- added input filter thread, that is responsible for actually receiving
  data (messages) from the CC IRQ
- the main thread now receives filtered data that is the average of the
  last _n_ messages

What I've essentially created is a rolling buffer, with one distinct
feature. If no messages are received in the configured lifespan, the
oldest messages are slowly removed from the queue & the sum/average is
also updated to reflect that.
This is to prevent cases where we stop receiving messages (stable or no
PWM) essentially leaving whatever the last _n_ messages were in the
buffer - as we want to reflect that the PWM has probably stabilized on
the last received value.

Hope everything makes sense, leave comments if it does not.